Move Sam(TypeAlias)ConstructorDescriptor to core
