+++ compiler/frontend.java/src/org/jetbrains/kotlin/load/java/sam/JavaSingleAbstractMethodUtils.java
@@ -28,9 +28,7 @@ import org.jetbrains.kotlin.load.java.descriptors.JavaMethodDescriptor;
 import org.jetbrains.kotlin.load.java.lazy.descriptors.LazyJavaClassDescriptor;
 import org.jetbrains.kotlin.name.Name;
 import org.jetbrains.kotlin.resolve.jvm.JavaResolverUtils;
-import org.jetbrains.kotlin.resolve.sam.SamConversionOracle;
-import org.jetbrains.kotlin.resolve.sam.SamConversionResolver;
-import org.jetbrains.kotlin.resolve.sam.SamConversionResolverImplKt;
+import org.jetbrains.kotlin.resolve.sam.*;
 import org.jetbrains.kotlin.types.KotlinType;
 import org.jetbrains.kotlin.types.SimpleType;
 import org.jetbrains.kotlin.types.TypeSubstitutor;
